+++ CotEditor/Sources/FormatPaneController.swift
@@ -514,6 +514,9 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
         alert.informativeText = "This action cannot be undone.".localized
         alert.addButton(withTitle: "Cancel".localized)
         alert.addButton(withTitle: "Delete".localized)
+        if #available(macOS 10.16, *) {
+            alert.buttons.last?.hasDestructiveAction = true
+        }
         
         let window = self.view.window!
         alert.beginSheetModal(for: window) { [unowned self] (returnCode: NSApplication.ModalResponse) in